Classification&Regression Models/Docker_Butterfly_Classification/requirements.txt
pandas
numpy
scikit-learn
Pillow
tqdm
matplotlib
joblib
uvicorn
python-multipart
fastapi
